requirements.txt
aiohttp
discord.py